CXF-8671: Support Jakarta EE 10 (#1201)
* Update dependencies to Jakarta EE 10

* Remove unnecessary dependency on jakarta.xml.bind-api

It didn’t respect the Jakarta EE 10 BOM for some reason, and it doesn’t (yet?) appear to be necessary

* Servlet API 6.0.0

* Jakarta XML Binding 4.0

* Jakarta Context Dependency Injection 4.0

* Fix Maven dependencies

* Jakarta RESTful Web Services 3.1

* Update Jakarta EE dependencies for codegen-plugins integration tests

* Add missing jakarta.xml.bind-api dependencies for codegen-plugin maven-invoker-plugin

* Correct RetryAfterHeaderProviderTest not accounting for host daylight time

getRawOffset() returns the offset for standard time, which means these tests fail when the default timezone is observing daylight time.

[ERROR] Failures:
[ERROR] org.apache.cxf.jaxrs.impl.RetryAfterHeaderProviderTest.test[retry after Sun Nov 03 08:49:37 2002 is duration of PT24884H49M37S from 2000-01-01 00:00:00.000]
[ERROR]   Run 1: RetryAfterHeaderProviderTest.test:73
Expected: is <PT24884H49M37S>
     but: was <PT24883H49M37S>
[ERROR]   Run 2: RetryAfterHeaderProviderTest.test:73
Expected: is <PT24884H49M37S>
     but: was <PT24883H49M37S>
[ERROR]   Run 3: RetryAfterHeaderProviderTest.test:73
Expected: is <PT24884H49M37S>
     but: was <PT24883H49M37S>
[ERROR]   Run 4: RetryAfterHeaderProviderTest.test:73
Expected: is <PT24884H49M37S>
     but: was <PT24883H49M37S>
[INFO]
[ERROR] org.apache.cxf.jaxrs.impl.RetryAfterHeaderProviderTest.test[retry after Sun Nov 3 08:49:37 2002 is duration of PT24884H49M37S from 2000-01-01 00:00:00.000]
[ERROR]   Run 1: RetryAfterHeaderProviderTest.test:73
Expected: is <PT24884H49M37S>
     but: was <PT24883H49M37S>
[ERROR]   Run 2: RetryAfterHeaderProviderTest.test:73
Expected: is <PT24884H49M37S>
     but: was <PT24883H49M37S>
[ERROR]   Run 3: RetryAfterHeaderProviderTest.test:73
Expected: is <PT24884H49M37S>
     but: was <PT24883H49M37S>
[ERROR]   Run 4: RetryAfterHeaderProviderTest.test:73
Expected: is <PT24884H49M37S>
     but: was <PT24883H49M37S>

* Upgrade saaj-impl

* Upgrade undertow

* Upgrade istack

* Placeholders for bootstrap API

* EntityPart

* Implement RuntimeDelegateImpl

* Partially implement TLS/SSL in SE bootstrap (needs some work on CXF engines)

* Fix cxf-tools-wadlto-jaxrs test cases

* Update Spring Boot to 3.1.0, Spring LDAP to 3.1.0, Spring Security to 6.1.0

* docs: Fix BUILDING.txt fastinstall line

* Fix compilation and checkstyle issues

* Update Hibernate to 6.2.9.Final

* Update Hibernate to 6.3.1.Final

* Refactor EntityPartBuilderImpl & EntityPartImpl

* Replace JakarteEE 10 BOM with indvidual component versions to simplify patching and minor updates

* Fixed CDI and Security test cases

* Update Weld to 5.1.2.Final

* Switch JSON-P/JSON-B implementation and fix test cases

---------

+++ rt/rs/security/oauth-parent/oauth2/src/main/java/org/apache/cxf/rs/security/oauth2/grants/code/JPACodeDataProvider.java
@@ -96,9 +96,11 @@ public class JPACodeDataProvider extends JPAOAuthDataProvider implements Authori
     }
 
     private ServerAuthorizationCodeGrant removeCodeGrant(String code, EntityManager em) throws OAuthServiceException {
-        ServerAuthorizationCodeGrant grant = em.getReference(ServerAuthorizationCodeGrant.class, code);
+        ServerAuthorizationCodeGrant grant = em.find(ServerAuthorizationCodeGrant.class, code);
         try {
-            em.remove(grant);
+            if (grant != null) {
+                em.remove(grant);
+            }
         } catch (EntityNotFoundException e) {
         }
         return grant;
